Base
included feature: rotation, gravity, velocity, angular velocity, flexibility in mass and shape of cube;
lots of functions that simplify the code used;
An object created by the rigidbody2d class has functions that can detect collisions with boxcolliders2D, right now it is only with the corners of the rigidbody. I will later the collision detection for the general edges.