organizing/re-making parts of code.
my physics equations are working. However, due to using gradients when returning the velocity value back into xspeed and yspeed it always ends up going in a positicve direction.
I will have to remake the section where the calculated speeds get put back intop the velocity, and make it based on the angle of the line to the world, and the angle of the wall to the world.